Remove duplicated etw_gecko files, move some additional functionality into shared
