+++ samply/src/windows/context_switch.rs
