--- samply/src/windows/etw_gecko/mod.rs
@@ -1,11 +1,0 @@
-mod context_switch;
-mod jit_category_manager;
-mod jit_function_add_marker;
-mod lib_mappings;
-mod marker_file;
-mod process_sample_data;
-mod stack_converter;
-mod stack_depth_limiting_frame_iter;
-mod timestamp_converter;
-mod types;
-mod unresolved_samples;
