+++ samply/src/windows/mod.rs
@@ -1,5 +1,6 @@
-mod etw;
 pub mod profiler;
+mod etw;
 mod winutils;
 mod etw_gecko;
 mod etw_reader;
+mod context_switch;
